Started work on the user card analytics activity and renamed a folder to work with currently implemented models. Also changed a drawable that I had been meaning to change in the logo for a while.
