+++ android/app/build.gradle
@@ -47,7 +47,7 @@ dependencies {
     implementation 'androidx.recyclerview:recyclerview:1.0.0'
 
     //Graphing Dependency
-    implementation 'com.jjoes64:graphview:4.2.2'
+    implementation 'com.jjoe64:graphview:4.2.2'
 }
 
 apply plugin: 'com.google.gms.google-services'
