+++ android/app/src/main/java/com/example/myfair/activities/ScannedActivity.java
@@ -7,12 +7,10 @@ import android.util.Log;
 import android.widget.TextView;
 
 
-import com.example.myfair.qrcode.EncryptionHelper;
-import com.example.myfair.qrcode.qrObject;
-
-import com.example.myfair.db.FirebaseDatabase;
 import com.example.myfair.modelsandhelpers.EncryptionHelper;
 import com.example.myfair.modelsandhelpers.qrObject;
+
+import com.example.myfair.db.FirebaseDatabase;
 
 import com.example.myfair.R;
 import com.example.myfair.db.Card;
