+++ android/app/src/main/java/com/example/myfair/activities/analytics/UserCardAnalyticsActivity.java
@@ -12,5 +12,6 @@ public class UserCardAnalyticsActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_user_card_analytics);
+
     }
 }
