+++ android/app/src/main/res/layout/activity_login.xml
@@ -45,7 +45,7 @@
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
             app:layout_constraintVertical_bias="0.155"
-            app:srcCompat="@drawable/icon_logo" />
+            app:srcCompat="@drawable/icon_logo_frame1" />
     </androidx.constraintlayout.widget.ConstraintLayout>
 
     <androidx.constraintlayout.widget.ConstraintLayout
