+++ android/app/src/main/res/values/strings.xml
@@ -65,5 +65,7 @@
 
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="graphTitle">Card Scans</string>
+    <string name="totalScansTitle">Total Scans</string>
 
 </resources>
